Alternative SM-based VE (to be removed)
An experimental, half-broken rewrite of VE Pulse to blink four strip sections to SM's marquee lighting output, also brings back marquee light fading (internally) and adds smoother fades on/off to it.

It's not a *terrible* effect in-game, but it's not great in menus and wouldn't look good if it was reading marquee lighting from non-stepmania games (say, if I read lighting output from a 573 - real DDR doesn't do anything really interesting).

Away it goes eventually, but committing so that it's at least saved somewhere before being nuked into oblivion.